Core: Count tests run so that 'suiteEnd' emits correctly with active filters
Previously, the `suiteEnd` event and the `moduleDone` callback
were fired when we've run all the tests, based on counting the
number of registered tests in a module and the number of tests
actually run.

This logic does not hold up when there are active module/test filters
used in the runner, because then some of the registered tests will
never run.

The added tests demonstrate the issue, as they will fail without
the src patch. Several of the `suiteEnd` events would not fire.

The impact of this bug was that the `suiteEnd` event and `moduleDone`
callback don't fire if there were active filters such as `--filters`
in the CLI or via the "Filter" field in the HTML Reporter.

This in turn meant that reporters such as TAP would sometimes
reflect the structure of the test suite in an odd way due to
it sometimes missing an "end" event for a group of tests.

Fixes https://github.com/qunitjs/qunit/issues/1416.
Closes https://github.com/qunitjs/qunit/pull/1417.